# openapi.yaml
openapi: 3.1.0
info:
  title: Payment & Report Service API
  version: 1.0.0
  description: |
    API for Payment & Report Service in the training center system. This service handles auth, user accounts, tuition fees, invoices, payments, and reports.
servers:
  - url: http://localhost:5203
    description: Local development server
paths:
...
  /api/payment-report/admin/users:
    get:
      summary: Get all users
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserAccount'
    post:
      summary: Create user account manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccount'
  /api/payment-report/admin/users/{userId}/status:
    patch:
      summary: Update user status
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
      responses:
        '200':
          description: Updated status
  /api/payment-report/admin/users/{userId}/reset-password:
    post:
      summary: Reset user password
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Password reset result
  /api/payment-report/admin/tuition-fees:
    get:
      summary: Get tuition fees
      responses:
        '200':
          description: Tuition fees list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TuitionFee'
    post:
      summary: Create tuition fee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TuitionFeeRequest'
      responses:
        '201':
          description: Created tuition fee
  /api/payment-report/admin/invoices:
    get:
      summary: Get invoices
      responses:
        '200':
          description: Invoice list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'
    post:
      summary: Create invoice manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequest'
      responses:
        '201':
          description: Created invoice
  /api/payment-report/admin/invoices/{invoiceId}/payments:
    post:
      summary: Record payment
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '201':
          description: Payment recorded
...
  /api/payment-report/internal/users/{userId}:
    get:
      summary: Get user by ID for internal use
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccount'
...
components:
  schemas:
...
    UserAccount:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        fullName:
          type: string
        role:
          type: string
        relatedEntityId:
          type: integer
        relatedEntityType:
          type: string
        status:
          type: string
        mustChangePassword:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        fullName:
          type: string
        role:
          type: string
        relatedEntityId:
          type: integer
        relatedEntityType:
          type: string
        mustChangePassword:
          type: boolean
      required:
        - username
        - email
        - role
        - relatedEntityType
    TuitionFee:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        classId:
          type: integer
          nullable: true
        amount:
          type: number
        currency:
          type: string
        effectiveFrom:
          type: string
          format: date
        isActive:
          type: boolean
    TuitionFeeRequest:
      type: object
      properties:
        courseId:
          type: integer
        classId:
          type: integer
          nullable: true
        amount:
          type: number
        currency:
          type: string
        effectiveFrom:
          type: string
          format: date
        isActive:
          type: boolean
      required:
        - courseId
        - amount
        - currency
        - effectiveFrom
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceCode:
          type: string
        studentId:
          type: integer
        enrollmentId:
          type: integer
        courseId:
          type: integer
        classId:
          type: integer
        totalAmount:
          type: number
        discountAmount:
          type: number
        finalAmount:
          type: number
        paidAmount:
          type: number
        debtAmount:
          type: number
        dueDate:
          type: string
          format: date
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateInvoiceRequest:
      type: object
      properties:
        studentId:
          type: integer
        enrollmentId:
          type: integer
        courseId:
          type: integer
        classId:
          type: integer
        totalAmount:
          type: number
        discountAmount:
          type: number
        finalAmount:
          type: number
        dueDate:
          type: string
          format: date
      required:
        - studentId
        - enrollmentId
        - courseId
        - finalAmount
        - dueDate
    PaymentRequest:
      type: object
      properties:
        amount:
          type: number
        paymentMethod:
          type: string
        collectedBy:
          type: string
        note:
          type: string
      required:
        - amount
        - paymentMethod
...